pyproject.toml
[project]
name = "hr-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "ipython>=9.6.0",
    "langchain>=0.3.27",
    "langchain-cli[inmem]>=0.0.37",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.37.0",
    "watchdog>=6.0.0",
]
